C++ redesign of data-structures
- Use flat arrays
- Use real type for all matrices
- Merge _f MaponiA3 files

+++ fmain.f90
@@ -6,7 +6,7 @@ program Interface_test
     integer i, j !! Iterators
     integer(c_int) :: dim, N_updates
     integer(c_int), dimension(:), allocatable :: Ar_index
-    integer(c_int), dimension(:,:), allocatable :: A, A0, Ar
+    real(c_double), dimension(:,:), allocatable :: A, A0, Ar
     real(c_double), dimension(:,:), allocatable :: A0_inv
 
     dim = 3
@@ -14,15 +14,15 @@ program Interface_test
     allocate(Ar_index(dim), A(dim,dim), A0(dim,dim), Ar(dim,dim), A0_inv(dim,dim))
 
     !! Initialize A with M=3 and fill acc. to Eq. (17) from paper
-    A(1,1) = 1
-    A(1,2) = 1
-    A(1,3) = -1
-    A(2,1) = 1
-    A(2,2) = 1
-    A(2,3) = 0
-    A(3,1) = -1
-    A(3,2) = 0
-    A(3,3) = -1
+    A(1,1) = 1.0d0
+    A(1,2) = 1.0d0
+    A(1,3) = -1.0d0
+    A(2,1) = 1.0d0
+    A(2,2) = 1.0d0
+    A(2,3) = 0.0d0
+    A(3,1) = -1.0d0
+    A(3,2) = 0.0d0
+    A(3,3) = -1.0d0
 
     !! Prepare the diagonal matrix A0 and the update matrix Ar
     do i=1,dim
@@ -32,13 +32,13 @@ program Interface_test
                 A0(i,j) = A(i,j)
                 A0_inv(i,j) = 1.0d0 / A0(i,j)
             else
-                A0(i,j) = 0
+                A0(i,j) = 0.0d0
                 A0_inv(i,j) = 0.0d0
             end if
             Ar(i,j) = A(i,j) - A0(i,j)
         end do
     end do
-    
+
     call MaponiA3(A0, A0_inv, dim, n_updates, Ar, Ar_index)
 
     do i=1,dim
